# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PaywayCheckout {
  success: boolean;
  tran_id: string;
  /**
     * @minLength 1
     * @maxLength 4096
     */
  qr_string: string;
  amount: string;
}

export interface PaywayStatusInput {
  /**
     * @minLength 1
     * @maxLength 128
     */
  tran_id: string;
}

export type PaywayPaymentStatusStatus = typeof PaywayPaymentStatusStatus[keyof typeof PaywayPaymentStatusStatus];


export const PaywayPaymentStatusStatus = {
  approved: 'approved',
  pending: 'pending',
  declined: 'declined',
  cancelled: 'cancelled',
  expired: 'expired',
  unknown: 'unknown',
} as const;

export interface PaywayPaymentStatus {
  tran_id: string;
  status: PaywayPaymentStatusStatus;
  message: string;
}

export interface ErrorResponse {
  error: string;
}