# lib/api-zod/src/generated/types/paywayCheckout.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaywayCheckout {
  success: boolean;
  tran_id: string;
  /**
     * @minLength 1
     * @maxLength 4096
     */
  qr_string: string;
  amount: string;
}