# lib/api-zod/src/generated/types/paywayPaymentStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaywayPaymentStatusStatus } from './paywayPaymentStatusStatus';

export interface PaywayPaymentStatus {
  tran_id: string;
  status: PaywayPaymentStatusStatus;
  message: string;
}