# lib/api-zod/src/generated/types/paywayStatusInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaywayStatusInput {
  /**
     * @minLength 1
     * @maxLength 128
     */
  tran_id: string;
}